feat(Help) Move flags package to pkg instead of under cmd
