+++ cmd/saml2aws/commands/configure.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/pkg/errors"
 	"github.com/versent/saml2aws"
-	"github.com/versent/saml2aws/cmd/saml2aws/commands/flags"
 	"github.com/versent/saml2aws/pkg/cfg"
+	"github.com/versent/saml2aws/pkg/flags"
 )
 
 // Configure configure account profiles
