+++ cmd/saml2aws/commands/exec.go
@@ -9,8 +9,8 @@ import (
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/sts"
 	"github.com/pkg/errors"
-	"github.com/versent/saml2aws/cmd/saml2aws/commands/flags"
 	"github.com/versent/saml2aws/pkg/awsconfig"
+	"github.com/versent/saml2aws/pkg/flags"
 	"github.com/versent/saml2aws/pkg/shell"
 )
