+++ cmd/saml2aws/commands/login_test.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/versent/saml2aws"
-	"github.com/versent/saml2aws/cmd/saml2aws/commands/flags"
 	"github.com/versent/saml2aws/pkg/cfg"
 	"github.com/versent/saml2aws/pkg/creds"
+	"github.com/versent/saml2aws/pkg/flags"
 )
 
 func TestResolveLoginDetailsWithFlags(t *testing.T) {
